use base class instead
