+++ src/Microsoft.SqlTools.LanguageService/Connection/Contracts/ConnectionDetails.cs
