+++ src/Microsoft.SqlTools.LanguageService/Connection/Contracts/ConnectionSummary.cs
